Add JAX services

+++ flowershop/src/main/java/com/accenture/flowershop/be/business/CartService.java
@@ -1,10 +1,11 @@
 package com.accenture.flowershop.be.business;
 
 import com.accenture.flowershop.be.entity.order.OrderItem;
+import java.math.BigDecimal;
 
 import java.util.List;
 
 public interface CartService {
     List<OrderItem> addFlowerToCart (List<OrderItem> cart, long flowerId, int count);
-    double getTotal(List<OrderItem> cart, int discount);
+    BigDecimal getTotal(List<OrderItem> cart, int discount);
 }
